src/templates/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
PyJWT==2.3.0
gunicorn==20.1.0
python-json-logger==2.0.7
requests==2.31.0
python-dateutil==2.8.2
structlog==23.1.0
flasgger